Move Application and related classes to Application namespace
