+++ ApplyToBecomeInternal/ApplyToBecome.Data/IApplications.cs
@@ -1,4 +1,4 @@
-using ApplyToBecome.Data.Models;
+using ApplyToBecome.Data.Models.Application;
 
 namespace ApplyToBecome.Data
 {
