+++ ApplyToBecomeInternal/ApplyToBecomeInternal.Tests/Models/ApplicationForm/Sections/AboutConversionSectionTests.cs
@@ -1,5 +1,5 @@
-﻿
-using ApplyToBecome.Data.Models;
+﻿using ApplyToBecome.Data.Models;
+using ApplyToBecome.Data.Models.Application;
 using ApplyToBecomeInternal.Models.ApplicationForm;
 using ApplyToBecomeInternal.Models.ApplicationForm.Sections;
 using FluentAssertions;
