+++ ApplyToBecomeInternal/ApplyToBecomeInternal.Tests/Models/ApplicationFormViewModelTests.cs
@@ -1,4 +1,5 @@
 using ApplyToBecome.Data.Models;
+using ApplyToBecome.Data.Models.Application;
 using ApplyToBecomeInternal.Models;
 using ApplyToBecomeInternal.Models.Shared;
 using FluentAssertions;
