+++ ApplyToBecomeInternal/ApplyToBecomeInternal/Models/ApplicationForm/Sections/AboutConversionSection.cs
@@ -1,6 +1,5 @@
-using ApplyToBecome.Data.Models;
+using ApplyToBecome.Data.Models.Application;
 using ApplyToBecomeInternal.Extensions;
-using System;
 using System.Collections.Generic;
 
 namespace ApplyToBecomeInternal.Models.ApplicationForm.Sections
