+++ ApplyToBecomeInternal/ApplyToBecomeInternal/Models/ApplicationForm/Sections/DeclarationSection.cs
@@ -1,4 +1,4 @@
-using ApplyToBecome.Data.Models;
+using ApplyToBecome.Data.Models.Application;
 
 namespace ApplyToBecomeInternal.Models.ApplicationForm.Sections
 {
